Remainder of files for refactor
